refactor: 유저 홈 컴포넌트 분리
